Simplify webpack.mix.js, move plugins to tasks/build.js
